feat(share-modal): promote 4 composition layouts to materials; fixes to atmosphere cards
Design set is now twelve materials in three families:
- artist (Dīwān, Ibn Muqla, Sinan, Zaha Hadid, Hassan Fathy)
- atmosphere (Layl, Mishkat, Sahifa)
- composition (Musnad, Muqabala, Najma, Iqtibas — promoted from drafts)

Per review:
- Wire Musnad, Muqabala, Najma, Iqtibas into shareCardDesigns.js as real
  selectable materials (registry, renderers, modal captions/accents, arch CSS).
- Remove the Neon ("burning glass") material entirely.
- Remove Layl's star; Layl is now pure minimal (no ornament).
- Move the book flourish from above the title to BETWEEN the header and the
  poem for the designs that carry it.
- Mishkat: the mihrab arch is now a wide border frame whose outline never
  crosses the title or verses; text is width-limited to the interior and the
  type steps down when many lines are selected.
- Sahifa: header and verses are pinned into the band between the masthead and
  colophon and shrink to fit, so even at six lines nothing rides over the top bar.

Shared canvas helpers added: fitFont (fit text to a max width) and drawBrandUnit
(alignment-safe brand mark). Drop the standalone layout-draft module and gallery;
the render-options gallery now shows all twelve materials.

Tests updated (12 designs; artist-only English-attribution check). 698 unit
tests pass; build and lint clean.
