+++ design-review/index.html
@@ -698,20 +698,10 @@ const CATALOG = [
     component: 'Share Modal',
     id: 'share-modal',
     categories: [
-      { id: 'share-renders', name: 'Card Renders — WYSIWYG', source: 'claude/design-3-4b-xru8t6', description: 'Post-3.4B implementation iteration — the preview IS the exported PNG (single canvas source of truth); nine card materials (five artist designs + four new directions), line selection up to six verse pairs, and centered/right alignment (all designs centre by default)',
+      { id: 'share-renders', name: 'Card Renders — WYSIWYG', source: 'claude/design-3-4b-xru8t6', description: 'Post-3.4B implementation — the preview IS the exported PNG (single canvas source of truth); twelve card materials across three families (artist, atmosphere, composition), line selection up to six verse pairs, and centered/right alignment (all designs centre by default)',
         options: [
-          { id: 'share-renders-gallery', name: 'Render Options Gallery', philosophy: 'What you see is exactly what you send — every material rendered from the same canvas as the export', file: 'share-modal/render-options.html', desc: 'All nine card renders at export resolution: the artist five (Dīwān, Ibn Muqla, Sinan, Zaha Hadid, Hassan Fathy) plus four new directions — Layl (midnight whisper), Mishkat (lantern niche), Sahifa (broadsheet), Neon (night sign)', generation: 10, iteration: 1, source_branch: 'claude/design-3-4b-xru8t6' },
-          { id: 'share-renders-flow', name: 'Flow — Lines & Alignment', philosophy: 'The reader chooses which verses travel and how the text is set', file: 'share-modal/render-options.html#flow', desc: 'WYSIWYG stage with floating card, glass lines panel (gold diamond checks per verse pair, up to six), centered/right alignment control, tap-outside to dismiss, dock-anchored toasts, live preview updates, material-accent chrome', generation: 10, iteration: 1, source_branch: 'claude/design-3-4b-xru8t6' }
-        ],
-        feedback: {}
-      },
-      { id: 'share-layout-drafts', name: 'Layout Drafts — Composition', source: 'claude/design-3-4b-xru8t6', description: 'Alternate card compositions holding colour constant (Dīwān palette) so the comparison is purely structural — layout, ornament, and how Arabic/English relate. Not yet wired into the live modal; a chosen direction gets promoted into the renderer',
-        options: [
-          { id: 'share-layout-musnad', name: 'Musnad — Numbered Margin', philosophy: 'The poem as an editorial manuscript page', file: 'share-modal/layout-drafts.html', desc: 'Poet as a top-left slug over a rule, verses numbered in Arabic-Indic numerals against a vertical margin rule; asymmetric with generous whitespace', generation: 10, iteration: 2, source_branch: 'claude/design-3-4b-xru8t6' },
-          { id: 'share-layout-matla', name: "Matla' — Hero Opening", philosophy: 'One line carries the card', file: 'share-modal/layout-drafts.html', desc: 'The opening hemistich set enormous and gold-glowing, a divider gem, then the remaining verses small beneath; pure scale hierarchy', generation: 10, iteration: 2, source_branch: 'claude/design-3-4b-xru8t6' },
-          { id: 'share-layout-najma', name: 'Najma — Star Medallion', philosophy: 'Verses cradled inside geometry', file: 'share-modal/layout-drafts.html', desc: 'An eight-point star (two overlapping rotated squares) frames the verses; title at the top point, poet at the bottom point; symmetric and ceremonial', generation: 10, iteration: 2, source_branch: 'claude/design-3-4b-xru8t6' },
-          { id: 'share-layout-muqabala', name: 'Muqabala — Facing Columns', philosophy: 'Parallel reading, not interleaved', file: 'share-modal/layout-drafts.html', desc: 'Arabic and English in two facing columns split by a central hairline rule with a gem at its midpoint; each column bounded so the languages never collide', generation: 10, iteration: 2, source_branch: 'claude/design-3-4b-xru8t6' },
-          { id: 'share-layout-iqtibas', name: 'Iqtibas — Pull-Quote', philosophy: 'The verse as a literary pull-quote', file: 'share-modal/layout-drafts.html', desc: 'An oversized low-opacity quotation mark top-left, a centred verse block, and attribution set on a short gold rule bottom-right; magazine feel', generation: 10, iteration: 2, source_branch: 'claude/design-3-4b-xru8t6' }
+          { id: 'share-renders-gallery', name: 'Render Options Gallery', philosophy: 'What you see is exactly what you send — every material rendered from the same canvas as the export', file: 'share-modal/render-options.html', desc: 'All twelve card renders at export resolution: the artist five (Dīwān, Ibn Muqla, Sinan, Zaha Hadid, Hassan Fathy); atmosphere (Layl, Mishkat, Sahifa); and composition layouts (Musnad numbered-margin, Muqabala facing-columns, Najma star-medallion, Iqtibas pull-quote)', generation: 11, iteration: 1, source_branch: 'claude/design-3-4b-xru8t6' },
+          { id: 'share-renders-flow', name: 'Flow — Lines & Alignment', philosophy: 'The reader chooses which verses travel and how the text is set', file: 'share-modal/render-options.html#flow', desc: 'WYSIWYG stage with floating card, glass lines panel (gold diamond checks per verse pair, up to six), centered/right alignment control, tap-outside to dismiss, dock-anchored toasts, live preview updates, material-accent chrome', generation: 11, iteration: 1, source_branch: 'claude/design-3-4b-xru8t6' }
         ],
         feedback: {}
       },
