--- design-review/share-modal/layout-drafts.html
@@ -1,86 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-<meta charset="UTF-8">
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
-<title>Share Card — Alternate Layout Drafts</title>
-<link rel="preconnect" href="https://fonts.googleapis.com">
-<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-<link href="https://fonts.googleapis.com/css2?family=Amiri:ital,wght@0,400;0,700&family=Reem+Kufi:wght@400;600;700&family=Playfair+Display:ital,wght@0,400;0,600;1,400&display=swap" rel="stylesheet">
-<style>
-:root { --gold:#c5a059; --gold-bright:#d4b463; --bg:#0c0c0e; --ink:rgba(232,224,208,0.92); }
-* { box-sizing:border-box; margin:0; padding:0; }
-body { background:var(--bg); color:var(--ink); font-family:system-ui,-apple-system,sans-serif; padding:48px 28px 90px; }
-header { max-width:1120px; margin:0 auto 8px; }
-.kicker { font-family:'Playfair Display',serif; font-style:italic; font-size:0.85rem; letter-spacing:0.14em; color:rgba(197,160,89,0.6); }
-h1 { font-family:'Playfair Display',serif; font-weight:600; font-size:1.7rem; margin:6px 0 4px; }
-h1 .ar { font-family:'Reem Kufi',sans-serif; color:var(--gold); font-size:1.2rem; margin-left:8px; }
-header p { color:rgba(232,224,208,0.62); font-size:0.92rem; max-width:70ch; line-height:1.65; }
-.rule { max-width:1120px; margin:26px auto; height:1px; background:linear-gradient(90deg,transparent,rgba(197,160,89,0.4),transparent); }
-.grid { max-width:1120px; margin:0 auto; display:grid; grid-template-columns:repeat(auto-fill,minmax(300px,1fr)); gap:26px; }
-.card { border:1px solid rgba(197,160,89,0.18); border-radius:16px; overflow:hidden; background:#101013; transition:transform .2s,border-color .2s; }
-.card:hover { transform:translateY(-4px); border-color:rgba(197,160,89,0.45); }
-.card img { display:block; width:100%; aspect-ratio:1080/1350; object-fit:cover; }
-figcaption { padding:14px 16px 18px; border-top:1px solid rgba(197,160,89,0.12); }
-.name { font-weight:600; font-size:1rem; display:flex; align-items:baseline; gap:8px; }
-.name .ar { font-family:'Reem Kufi',sans-serif; font-size:0.88rem; color:var(--gold); }
-.tag { margin-left:auto; font-size:0.6rem; font-weight:700; letter-spacing:0.12em; color:#191307; background:linear-gradient(135deg,#b8922e,#e2c67a); border-radius:999px; padding:2px 9px; }
-.desc { margin-top:6px; font-family:'Playfair Display',serif; font-style:italic; font-size:0.82rem; color:rgba(197,160,89,0.68); line-height:1.5; }
-.note { max-width:1120px; margin:26px auto 0; font-size:0.82rem; color:rgba(232,224,208,0.5); line-height:1.6; }
-</style>
-</head>
-<body>
-<header>
-  <div class="kicker">SHARE MODAL — ALTERNATE LAYOUT STUDIES</div>
-  <h1>Card Layout Drafts <span class="ar">تخطيطات</span></h1>
-  <p>
-    Five alternate <em>compositions</em> for the share card. Colour is held constant
-    (the Dīwān obsidian-and-gold palette) on purpose — so the comparison is purely about
-    structure: how the poem is arranged, where the ornament lives, how Arabic and English
-    relate on the page. Pick a direction and it gets promoted into the real renderer.
-  </p>
-</header>
-<div class="rule"></div>
-<div class="grid">
-  <figure class="card">
-    <img src="layouts/musnad.png" alt="Musnad layout" loading="lazy">
-    <figcaption>
-      <div class="name">Musnad <span class="ar">مسند</span> <span class="tag">DRAFT</span></div>
-      <div class="desc">Numbered margin — editorial manuscript. Poet as a top-left slug, verses numbered against a vertical margin rule. Asymmetric, lots of quiet.</div>
-    </figcaption>
-  </figure>
-  <figure class="card">
-    <img src="layouts/matla.png" alt="Matla layout" loading="lazy">
-    <figcaption>
-      <div class="name">Matla' <span class="ar">مطلع</span> <span class="tag">DRAFT</span></div>
-      <div class="desc">Hero opening — the first hemistich set enormous and glowing; the rest follow small. One line is the star.</div>
-    </figcaption>
-  </figure>
-  <figure class="card">
-    <img src="layouts/najma.png" alt="Najma layout" loading="lazy">
-    <figcaption>
-      <div class="name">Najma <span class="ar">نجمة</span> <span class="tag">DRAFT</span></div>
-      <div class="desc">Star medallion — verses cradled inside an eight-point geometric star. Symmetric, ceremonial.</div>
-    </figcaption>
-  </figure>
-  <figure class="card">
-    <img src="layouts/muqabala.png" alt="Muqabala layout" loading="lazy">
-    <figcaption>
-      <div class="name">Muqabala <span class="ar">مقابلة</span> <span class="tag">DRAFT</span></div>
-      <div class="desc">Facing columns — Arabic and English side by side, split by a central rule. Parallel reading instead of interleaved.</div>
-    </figcaption>
-  </figure>
-  <figure class="card">
-    <img src="layouts/iqtibas.png" alt="Iqtibas layout" loading="lazy">
-    <figcaption>
-      <div class="name">Iqtibas <span class="ar">اقتباس</span> <span class="tag">DRAFT</span></div>
-      <div class="desc">Pull-quote — an oversized quotation motif, literary block, attribution on a gold rule bottom-right. Magazine feel.</div>
-    </figcaption>
-  </figure>
-</div>
-<p class="note">
-  Drafts rendered from <code>src/utils/shareCardLayouts.draft.js</code> (canvas, export-resolution 1080×1350) —
-  not yet wired into the live modal. Each honours line-selection and could adopt the centre/right alignment toggle.
-</p>
-</body>
-</html>
